requirements.txt
appdirs==1.4.3
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
ConfigArgParse==0.14.0
datrie==0.8
docutils==0.15.2
fire==0.2.1
gitdb2==2.0.6
GitPython==3.0.2
idna==2.8
jsonschema==3.0.2
pkg-resources==0.0.0
pyrsistent==0.15.4
python-dotenv==0.10.3
PyYAML==5.1.2
ratelimiter==1.2.0.post0
requests==2.22.0
six==1.12.0
smmap2==2.0.5
snakemake==5.6.0
termcolor==1.1.0
tqdm==4.36.1
urllib3==1.25.6
wrapt==1.11.2
appdirs==1.4.3
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
ConfigArgParse==0.14.0
dask==2.5.0
datrie==0.8
docutils==0.15.2
fire==0.2.1
gitdb2==2.0.6
GitPython==3.0.2
idna==2.8
jsonschema==3.0.2
pkg-resources==0.0.0
pyrsistent==0.15.4
python-dotenv==0.10.3
PyYAML==5.1.2
ratelimiter==1.2.0.post0
requests==2.22.0
six==1.12.0
smmap2==2.0.5
snakemake==5.6.0
termcolor==1.1.0
toolz==0.10.0
tqdm==4.36.1
urllib3==1.25.6
wrapt==1.11.2
